+++ frontend/package.json
@@ -13,6 +13,7 @@
     "react-router-dom": "^6.22.3",
     "react-scripts": "^5.0.1",
     "tailwindcss": "^3.4.1",
+    "typescript": "^5.4.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
